add Object::wrap() method.
We can now re-wrap a tree after an object is created. This is useful
e.g. when you would like to have an Object member var that's multiple
layers deep below an ancerstor object that's passed into your ctor --
yuo can default construct this descendant object and then traverse the
ancestor until you get to the Object you wish to connect to.